requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorlog==6.9.0
cryptography==44.0.2
distro==1.9.0
ecdsa==0.19.1
Faker==37.1.0
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.75.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0
